+++ dist2/src/data-model/excel-table.d.ts
@@ -25,6 +25,8 @@ export interface Header {
     label: string;
     text: string;
     size?: number;
+    multiStyleValue?: MultiStyleValue;
+    comment?: Comment | string;
     formula?: {
         type: FormulaType;
         styleId?: string;
@@ -33,12 +35,33 @@ export interface Header {
 export interface Data extends DataOptions {
     [key: string]: string | number | any | undefined;
 }
+export interface MultiStyleRexValue {
+    reg: RegExp | string;
+    styleId: string;
+}
+export interface Comment {
+    comment?: string;
+    styleId?: string;
+    author?: string;
+}
+export interface MultiStyleValue {
+    [key: string]: string | undefined | MultiStyleRexValue[];
+    reg?: MultiStyleRexValue[];
+}
+export interface MapMultiStyleValue {
+    [key: string]: MultiStyleValue;
+}
+export interface MapComment {
+    [key: string]: Comment | string;
+}
 export interface DataOptions {
-    [key: string]: "0" | "1" | number | string | undefined;
+    [key: string]: "0" | "1" | number | string | undefined | MapComment | MapMultiStyleValue;
     outlineLevel?: number;
     hidden?: "0" | "1" | number;
     rowStyle?: string;
     height?: number;
+    multiStyleValue?: MapMultiStyleValue;
+    comment?: MapComment;
 }
 export interface MergeRowConditionMap {
     [columnKey: string]: {
@@ -60,6 +83,8 @@ export interface Title {
     height?: number;
     styleId?: string;
     text?: string;
+    multiStyleValue?: MultiStyleValue;
+    comment?: Comment | string;
 }
 export interface Sheet {
     withoutHeader?: boolean;
@@ -86,6 +111,11 @@ export interface HeaderRowOption {
     outlineLevel: "string";
 }
 export interface StyleMapper {
+    commentSintax: {
+        value: {
+            [key: string]: string;
+        };
+    };
     format: {
         count: number;
         value: string;