+++ dist2/src/utils/content-generator/content-types.d.ts
@@ -1,2 +1,2 @@
-export declare function contentTypeGenerator(sheetContentType: string): string;
+export declare function contentTypeGenerator(sheetContentType: string, commentId: number[]): string;
 //# sourceMappingURL=content-types.d.ts.map