+++ dist2/src/utils/multi-value.d.ts
@@ -1,0 +1,5 @@
+import { MultiStyleValue } from "../data-model/excel-table";
+export declare function generateMultiStyleValue(multiStyle: MultiStyleValue, text: string, styles: {
+    [key: string]: string;
+}): string;
+//# sourceMappingURL=multi-value.d.ts.map